server/requirements.txt
Flask
google-generativeai
Werkzeug 